@@ -1,8 +1,8 @@
-// +build tools
+//go:build tools
 
 package tools
 
 import (
-	// document generation
+	// Documentation generation
 	_ "github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs"
 )
