@@ -2,10 +2,9 @@ package main
 
 import (
 	"context"
-	"flag"
 	"log"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
+	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/netascode/terraform-provider-nxos/internal/provider"
 )
 
@@ -29,20 +28,13 @@ var (
 )
 
 func main() {
-	var debugMode bool
-
-	flag.BoolVar(&debugMode, "debug", false, "set to true to run the provider with support for debuggers like delve")
-	flag.Parse()
-
-	opts := &plugin.ServeOpts{ProviderFunc: provider.New(version)}
-
-	if debugMode {
-		err := plugin.Debug(context.Background(), "registry.terraform.io/netascode/nxos", opts)
-		if err != nil {
-			log.Fatal(err.Error())
-		}
-		return
+	opts := tfsdk.ServeOpts{
+		Name: "registry.terraform.io/netascode/nxos",
 	}
 
-	plugin.Serve(opts)
+	err := tfsdk.Serve(context.Background(), provider.New(version), opts)
+
+	if err != nil {
+		log.Fatal(err.Error())
+	}
 }
